move pub key store to keystores dir
