+++ src/backingServices/mongo.spec.ts
@@ -2,7 +2,7 @@ import { EnvVarError } from 'env-var';
 import mongoose, { Connection } from 'mongoose';
 
 import { mockSpy, MONGO_ENV_VARS } from '../_test_utils';
-import { MongoPublicKeyStore } from '../MongoPublicKeyStore';
+import { MongoPublicKeyStore } from '../keystores/MongoPublicKeyStore';
 import { configureMockEnvVars } from '../services/_test_utils';
 import {
   createMongooseConnectionFromEnv,
